+++ source/extensions/omni.isaac.lab/omni/isaac/lab/controllers/operational_space.py
@@ -77,11 +77,15 @@ class OperationalSpaceController:
         self._selection_matrix_force_b = torch.zeros_like(self._selection_matrix_force_task)
         # -- commands
         self._task_space_target_task = torch.zeros(self.num_envs, self.target_dim, device=self._device)
-        # -- buffers for motion/force control
+        # -- Placeholders for motion/force control
         self.desired_ee_pose_task = None
         self.desired_ee_pose_b = None
         self.desired_ee_wrench_task = None
         self.desired_ee_wrench_b = None
+        # -- buffer for operational space mass matrix
+        self._os_mass_matrix_b = torch.zeros(self.num_envs, 6, 6, device=self._device)
+        # -- Placeholder for the inverse of joint space mass matrix
+        self._mass_matrix_inv = None
         # -- motion control gains
         self._motion_p_gains_task = torch.diag_embed(
             torch.ones(self.num_envs, 6, device=self._device)
@@ -126,6 +130,14 @@ class OperationalSpaceController:
         )
         # -- end-effector contact wrench
         self._ee_contact_wrench_b = torch.zeros(self.num_envs, 6, device=self._device)
+
+        # -- buffers for null-space control gains
+        self._nullspace_p_gain = torch.tensor(self.cfg.nullspace_stiffness, dtype=torch.float, device=self._device)
+        self._nullspace_d_gain = (
+            2
+            * torch.sqrt(self._nullspace_p_gain)
+            * torch.tensor(self.cfg.nullspace_damping_ratio, dtype=torch.float, device=self._device)
+        )
 
     """
     Properties.
@@ -338,6 +350,9 @@ class OperationalSpaceController:
         current_ee_force_b: torch.Tensor | None = None,
         mass_matrix: torch.Tensor | None = None,
         gravity: torch.Tensor | None = None,
+        current_joint_pos: torch.Tensor | None = None,
+        current_joint_vel: torch.Tensor | None = None,
+        nullspace_joint_pos_target: torch.Tensor | None = None,
     ) -> torch.Tensor:
         """Performs inference with the controller.
 
@@ -348,12 +363,18 @@ class OperationalSpaceController:
                 (``num_envs``, 7), which contains the position and quaternion ``(w, x, y, z)``. Defaults to ``None``.
             current_ee_vel_b: The current end-effector velocity in root frame. It is a tensor of shape
                 (``num_envs``, 6), which contains the linear and angular velocities. Defaults to None.
-            current_ee_force_b: The current external force on the end-effector in root frame.
-                It is a tensor of shape (``num_envs``, 3), which contains the linear force. Defaults to ``None``.
+            current_ee_force_b: The current external force on the end-effector in root frame. It is a tensor of
+                shape (``num_envs``, 3), which contains the linear force. Defaults to ``None``.
             mass_matrix: The joint-space mass/inertia matrix. It is a tensor of shape (``num_envs``, ``num_DoF``,
-            ``num_DoF``). Defaults to ``None``.
+                ``num_DoF``). Defaults to ``None``.
             gravity: The joint-space gravity vector. It is a tensor of shape (``num_envs``, ``num_DoF``). Defaults
-            to ``None``.
+                to ``None``.
+            current_joint_pos: The current joint positions. It is a tensor of shape (``num_envs``, ``num_DoF``).
+                Defaults to ``None``.
+            current_joint_vel: The current joint velocities. It is a tensor of shape (``num_envs``, ``num_DoF``).
+                Defaults to ``None``.
+            nullspace_joint_pos_target: The target joint positions the null space controller is trying to enforce, when
+                possible. It is a tensor of shape (``num_envs``, ``num_DoF``).
 
         Raises:
             ValueError: When motion-control is enabled but the current end-effector pose or velocity is not provided.
@@ -361,6 +382,14 @@ class OperationalSpaceController:
             ValueError: When the current end-effector pose is not provided for the ``pose_rel`` command.
             ValueError: When closed-loop force control is enabled but the current end-effector force is not provided.
             ValueError: When gravity compensation is enabled but the gravity vector is not provided.
+            ValueError: When null-space control is enabled but the system is not redundant.
+            ValueError: When dynamically consistent pseudo-inverse is enabled but the mass matrix inverse is not
+                provided.
+            ValueError: When null-space control is enabled but the current joint positions and velocities are not
+                provided.
+            ValueError: When target joint positions are provided for null-space control but their dimensions do not
+                match the current joint positions.
+            ValueError: When an invalid null-space control method is provided.
 
         Returns:
             Tensor: The joint efforts computed by the controller. It is a tensor of shape (``num_envs``, ``num_DoF``).
@@ -398,23 +427,21 @@ class OperationalSpaceController:
                 if mass_matrix is None:
                     raise ValueError("Mass matrix is required for inertial decoupling.")
                 # Compute operational space mass matrix
-                mass_matrix_inv = torch.inverse(mass_matrix)
+                self._mass_matrix_inv = torch.inverse(mass_matrix)
                 if self.cfg.partial_inertial_dynamics_decoupling:
-                    # Create a zero tensor representing the mass matrix, to fill in the non-zero elements later
-                    os_mass_matrix = torch.zeros(self.num_envs, 6, 6, device=self._device)
                     # Fill in the translational and rotational parts of the inertia separately, ignoring their coupling
-                    os_mass_matrix[:, 0:3, 0:3] = torch.inverse(
-                        jacobian_b[:, 0:3] @ mass_matrix_inv @ jacobian_b[:, 0:3].mT
+                    self._os_mass_matrix_b[:, 0:3, 0:3] = torch.inverse(
+                        jacobian_b[:, 0:3] @ self._mass_matrix_inv @ jacobian_b[:, 0:3].mT
                     )
-                    os_mass_matrix[:, 3:6, 3:6] = torch.inverse(
-                        jacobian_b[:, 3:6] @ mass_matrix_inv @ jacobian_b[:, 3:6].mT
+                    self._os_mass_matrix_b[:, 3:6, 3:6] = torch.inverse(
+                        jacobian_b[:, 3:6] @ self._mass_matrix_inv @ jacobian_b[:, 3:6].mT
                     )
                 else:
                     # Calculate the operational space mass matrix fully accounting for the couplings
-                    os_mass_matrix = torch.inverse(jacobian_b @ mass_matrix_inv @ jacobian_b.mT)
+                    self._os_mass_matrix_b[:] = torch.inverse(jacobian_b @ self._mass_matrix_inv @ jacobian_b.mT)
                 # (Generalized) operational space command forces
                 # F = (J M^(-1) J^T)^(-1) * \ddot(x_des) = M_task * \ddot(x_des)
-                os_command_forces_b = os_mass_matrix @ des_ee_acc_b
+                os_command_forces_b = self._os_mass_matrix_b @ des_ee_acc_b
             else:
                 # Task-space impedance control: command forces = \ddot(x_des).
                 # Please note that the definition of task-space impedance control varies in literature.
@@ -454,5 +481,68 @@ class OperationalSpaceController:
                 raise ValueError("Gravity vector is required for gravity compensation.")
             # add gravity compensation
             joint_efforts += gravity
+
+        # Add null-space control
+        # -- Free null-space control
+        if self.cfg.nullspace_control == "none":
+            # No additional control is applied in the null space.
+            pass
+        else:
+            # Check if the system is redundant
+            if num_DoF <= 6:
+                raise ValueError("Null-space control is only applicable for redundant manipulators.")
+
+            # Calculate the pseudo-inverse of the Jacobian
+            if self.cfg.inertial_dynamics_decoupling and not self.cfg.partial_inertial_dynamics_decoupling:
+                # Dynamically consistent pseudo-inverse allows decoupling of null space and task space
+                if self._mass_matrix_inv is None or mass_matrix is None:
+                    raise ValueError("Mass matrix inverse is required for dynamically consistent pseudo-inverse")
+                jacobian_pinv_transpose = self._os_mass_matrix_b @ jacobian_b @ self._mass_matrix_inv
+            else:
+                # Moore-Penrose pseudo-inverse if full inertia matrix is not available (e.g., no/partial decoupling)
+                jacobian_pinv_transpose = torch.pinverse(jacobian_b).mT
+
+            # Calculate the null-space projector
+            nullspace_jacobian_transpose = (
+                torch.eye(n=num_DoF, device=self._device) - jacobian_b.mT @ jacobian_pinv_transpose
+            )
+
+            # Null space position control
+            if self.cfg.nullspace_control == "position":
+
+                # Check if the current joint positions and velocities are provided
+                if current_joint_pos is None or current_joint_vel is None:
+                    raise ValueError("Current joint positions and velocities are required for null-space control.")
+
+                # Calculate the joint errors for nullspace position control
+                if nullspace_joint_pos_target is None:
+                    nullspace_joint_pos_target = torch.zeros_like(current_joint_pos)
+                # Check if the dimensions of the target nullspace joint positions match the current joint positions
+                elif nullspace_joint_pos_target.shape != current_joint_pos.shape:
+                    raise ValueError(
+                        f"The target nullspace joint positions shape '{nullspace_joint_pos_target.shape}' does not"
+                        f"match the current joint positions shape '{current_joint_pos.shape}'."
+                    )
+
+                joint_pos_error_nullspace = nullspace_joint_pos_target - current_joint_pos
+                joint_vel_error_nullspace = -current_joint_vel
+
+                # Calculate the desired joint accelerations
+                joint_acc_nullspace = (
+                    self._nullspace_p_gain * joint_pos_error_nullspace
+                    + self._nullspace_d_gain * joint_vel_error_nullspace
+                ).unsqueeze(-1)
+
+                # Calculate the projected torques in null-space
+                if mass_matrix is not None:
+                    tau_null = (nullspace_jacobian_transpose @ mass_matrix @ joint_acc_nullspace).squeeze(-1)
+                else:
+                    tau_null = nullspace_jacobian_transpose @ joint_acc_nullspace
+
+                # Add the null-space joint efforts to the total joint efforts
+                joint_efforts += tau_null
+
+            else:
+                raise ValueError(f"Invalid null-space control method: {self.cfg.nullspace_control}.")
 
         return joint_efforts