+++ source/extensions/omni.isaac.lab_tasks/docs/CHANGELOG.rst
@@ -1,6 +1,17 @@
 Changelog
 ---------
 
+0.10.17 (2024-12-17)
+~~~~~~~~~~~~~~~~~~~~
+
+Changed
+^^^^^^^
+
+* Changed the configuration of
+  :class:`omni.isaac.lab.envs.mdp.actions.OperationalSpaceControllerAction`
+  inside the ``Isaac-Reach-Franka-OSC-v0`` environment to enable nullspace control.
+
+
 0.10.16 (2024-12-16)
 ~~~~~~~~~~~~~~~~~~~~